Fix Linux CI prerequisites and asset-free test execution
Probe development packages independently of the compiler and Vulkan runtime,
including the Autotools required by vcpkg's X11 dependencies. Exercise this
preflight with five Python regression tests on both hosted platforms and
select GCC 14 explicitly for the GCC and Linux lint jobs.

Resolve asset fixtures before Catch2 assertions: SKIP thrown from inside
REQUIRE is treated as a failed assertion, not a skipped test. Keep every
asset-backed assertion active when the data is present.

Clear exposed narrowing, aggregate-default, unused-code and include diagnostics
without weakening the lint gate. Ignore generated Python bytecode.

Validated: 540 local tests with assets, 540 asset-free tests with expected
missing-data skips, setup tests on Windows and Ubuntu, and 120-frame normal
and title smoke runs. clangd's 442-file sweep and focused correction recheck
are clean. Windows LLVM 18's analyzer crashes on existing WorldDataWadTests;
Linux CI remains the authoritative full clang-tidy check.

+++ src/formats/ObjWriter.cpp
@@ -12,11 +12,12 @@ std::string encodeObj(const Mesh& mesh, std::string_view name) {
     for (const MeshVertex& v : mesh.vertices) {
         if (mesh.prelit) {
             out += std::format("v {:.6g} {:.6g} {:.6g} {:.4g} {:.4g} {:.4g}\n", v.position.x,
-                               v.position.y, v.position.z, v.color.r / 255.0f, v.color.g / 255.0f,
-                               v.color.b / 255.0f);
+                               v.position.y, v.position.z, static_cast<f32>(v.color.r) / 255.0f,
+                               static_cast<f32>(v.color.g) / 255.0f,
+                               static_cast<f32>(v.color.b) / 255.0f);
         } else {
-            out += std::format("v {:.6g} {:.6g} {:.6g}\n", v.position.x, v.position.y,
-                               v.position.z);
+            out +=
+                std::format("v {:.6g} {:.6g} {:.6g}\n", v.position.x, v.position.y, v.position.z);
         }
     }
     for (const MeshVertex& v : mesh.vertices) {
